// src/interfaces/Planet.ts
export interface Planet {
  name: string
  rotation_period: string
  orbital_period: string
  diameter: string
  climate: string
  gravity: string
  terrain: string
  surface_water: string
  population: string
  residents: string[]
  films: string[]
  created: Date
  edited: Date
  url: string
}

export interface PlanetsRequest {
  count: number
  next: string
  previous?: any
  results: Planet[]
}